Add bin/summarize-clade-sequence-counts
Takes a metadata TSV and creates a summary of sequence counts grouped
by date, location, and clade.

The metadata can be filtered prior to generating the summary counts by
using the `--filter-query` option inspired by the `--query` option
of `augur filter`.

The `--filter-columns` must be provided along with the `--filter-query`
in order to load the correct columns from the metadata TSV. This
results in faster parsing time and lower memory usage.