fetch-ncov-us-case-counts: Add total count limit
First check the total number of records for the CDC dataset then fetch
the data with the total count set as the limit parameter. This is
required since the default limit for all SODA API calls is 1000 but
there's no way to ask for all records¹.

I'm still choosing to use the API over the direct CSV download link
since it allows us to cut and filter records before the data transfer
and the dates returned from the API are already in ISO 8601 format.

¹ https://dev.socrata.com/docs/paging.html