Add .gitignore
Ignore files in `data/` directory since this will be used to house the
subset of data used for analysis.

Ignore `logs/` directory since this will be used to house logs from the
pipeline.

Ignore `*.err` files, which are error outputs from Cram tests.

Ignore `.snakemake/` directory, which are outputs for local Snakemake
runs that do not need to be in version control.