Add scripts/prepare-data.py
Add script used to prepare the clades and cases data for analysis.
This script includes options to allow:
1. Subsetting data by cutoff date + days leading up to the cutoff
2. Only including locations that have a minimum number of sequences over
a specific number of days leading to the cutoff date
3. Excluding data with a manually annotated list of locations
4. Excluding recent sequence counts in a specific number of days leading
to the cutoff date
5. Collapsing rare clades into 'other' by giving a minimum number of
sequences per clade in a specific number of days leading to the cutoff.

These are all available as options to make it easy to tweak how we want
to subsample the entire dataset. However, the column names and date
format have been hardcoded as this script is only expected to work with
the output files from the ingest pipeline.

The script prints out the data subset rules as it processes the options
so that we can have a final log output of data prep rules.