Add default locations to exclude
Include a list of countries that we want to exclude from our global
analysis. This list was pulled from the most recent version of the
blab/omicron-countries-split data prep on 22 June 2022¹

Include a list of states that we want to exclude from our USA analysis.
This list was pulled from the most recent version of the
blab/omicron-us-split data prep on 22 June 2022²

¹ https://github.com/blab/rt-from-frequency-dynamics/blob/8a66fb75bcbdcd9c19245c6d4ca9877933253c49/data/omicron-countries-split/omicron-countries-split_data-prep.nb#L364
² https://github.com/blab/rt-from-frequency-dynamics/blob/68183a652fbb59980a3899766df17c2fd9734496/data/omicron-us-split/omicron-us-split_data-prep.nb#L2217