Add Snakefile
Only includes the data download and data preparation rules since it is
not clear yet how we want to set up the config for running `evofr`.
The default output for now is the outputs from the data preparation.

Added wildcard constraint for `geo_resolution` in the data download
rules to ensure that we only try to pull data that have been created by
the ingest pipeline.

Config must include the `data_provenances` and `geo_resolutions` params
so that we know which data to pull. All of the options for the
prepare_data.py script are optional, so they are optional params to
include in the config.