Add Cram tests for `prepare-data.py`
Cram tests are named with a numeric prefix to show the order of
compounding options used in the tests. The tests uses `tsv-utils` to
slice the outputs to do a high level check that they include the
expected locations, clades, and date ranges.

The testing data is a small subset of the public global clade and case
counts.The test data dates range from 2022-01-01 to 2022-01-10 to
include a variety of clades. I chose Argentina, Japan, USA, and
United Kingdom as the locations since they represent various levels of
sequencing in the public data. The example excluded_locations.txt is
just for testing purposes.