prepare-data: Change `clade` to `variant`
In the `evofr` package, we are expected to have a `variant` column
instead of a `clade` column. Add the `variant` column, which is
currently just a copy of the `clade` column. Subsequent commits will add
a default mapper of clades to variants since this is not exactly a 1:1
relationship. The collapsing of rare clades to an "other" group has also
been updated to change the `variant` value instead of the `clade` value.

Any references to "clade" after the data prep step have been changed to
"variant" to reflect this change.

The Cram tests for `prepare-data' have also been updated the reflect
this change.