prepare-data: Add option for clade-to-variant map
Since the mapping of clade to variants is not 1:1, we can provide
a file that is a mapping of clade names to variant names. Clades that
are not included in the mapping are automatically added to the "other"
variant group.

This commit includes the default map of clades to variants, which
follows the mapping that we discussed internally on Slack¹.

The Cram tests have been updated to use a clade-to-variant map that just
uses the clade name as the variant name so that the content checks do
not have to be updated.

¹ https://bedfordlab.slack.com/archives/C036L7V615F/p1659651323516289?thread_ts=1659554446.598349&cid=C036L7V615F