prepare-data: Add option to output clade without variant
Prior to this commit, clades without matching variants were just
silently assigned to the "other" variant group. This commit adds the
option to output these clades to a file before assigning them to the
"other" group. A new Cram test has been added to test this new option.

In a later commit, I'll add the snakemake rule to send this
file to a Slack message so that we can be alerted about these clades.