prepare-data: Add option to force include clades
Add option to force include clades in prepared data so that we have a
way to bypass the sequences counts for specific clades. This is also a
way to override the default clade to variant mapping since you must
provide the clade name and its corresponding variant name. A new Cram
test has been added ot test this new option.