Update model scripts to take override input/output
Allow model scripts to take command line args for input and output paths
so that it's easy to slot these scripts into the Snakemake pipeline.

These command line args will just override the input and output paths
provided in the model config files, so it is still possible to run the
model scripts with only a config YAML file.