scripts/prepare-data: specify `numeric_only=False`
Starting with pandas v1.5.0, we get a deprecation warning for using
the default value `numeric_only=None` in `sum()`, which led to cram
tests failing.

This commit updates to use the new suggested value `numeric_only=False`.
This change requires us to explicitly drop the "clade" column, which is
better and less confusing anyways!