Update model scripts to take override data name
Override the config data name with command line options so that we are
able to define the results filename in the Snakemake pipeline without
parsing the YAML configs.

The subsequent commit will use this option to make the data name the run
date so that we can have dated result files.