renewal config: Add generation time for new variants
Added 22D, 22E, 22F in the clade_to_variant map in 4c16e98, but I forgot
to add their generation times to the renewal config. This resulted in
the following error when running the renewal model:

> AssertionError: All variants must be present in config or have same gen time.

Confirmed with @marlinfiggins on Slack¹ that the new variants should
have the same mean and sd as the other Omicron variants.

¹ https://bedfordlab.slack.com/archives/C036L7V615F/p1669077133665769?thread_ts=1669074666.216229&cid=C036L7V615F