Replace fetch of counts files with rules to create them
Previously, case and clade counts files were generated by upstream
GH Action workflows that uploads them to S3. This commit consolidates
those steps into the main pipeline so that the counts files are
generated at the start of pipeline instead of fetching the files
from S3.

This makes it easier to track when the raw data files were fetched from
external sources and when the case/clade counts were created.