Add rules for uploads of case and clade counts
Adds steps from the GH Action workflows to upload counts files as rules
in the main Snakemake pipeline.

Note the case and clades uploads are kept as separate rules because
I am anticipating the addition of a "clade" wildcard to the clades
upload inputs/outputs as we expand the pipeline to include other clades
such as Nextclade pangolin.