[viz] Add legend
The intention is for a right-hand-sidebar style legend on wide
screens and an inline sidebar for smaller screens. There is still some
work to do for smaller screens where we want to duplicate the legend
for each set of small-multiples. We may also need to revisit this if we
require multiple (different) legends on a single page.