ingest: Add clade type to S3 prefix
We are planning on adding pango lineage clade counts to this file
structure in the near future, so it makes sense to include the clade
type as a prefix for the clade counts files. Each clade type is split
into clade counts by geographic resolution, so the clade type is part of
the prefix while the geo-resolution is the file name.