Add rule upload_model_results_to_s3_as_latest
Uploads the dated model results to S3 as `latest_results.json.gz` so
that the visualization app can easily grab the latest results.

I added this as a separate rule instead of adding as second upload to
the existing rule so that they can run in parallel. There should be a
way to consolidate the two upload rules, but I didn't want to think
through the wildcards manipulation to do this.