Use Nextstrain CLI to upload model results
Nextstrain CLI does the right thing of uploading gzip JSON files with
the correct content-encoding and content-type, which are necessary for
the visualization app to properly decompress the model results.

Added this as a separate bash script so that we can continue to send
Slack notifications after the files have been uploaded.

Removed the `cloudfront_domain` param from the optional config since
the Nextstrain CLI automatically does the CloudFront invalidation.

Updated README to reflect that the uploaded files will no longer have
the `.gz` suffix because of the switch to uploading with the Nextstrain
CLI.