Add generate-static-model-viz GH workflow
Uses the viz app and Puppeteer to generate static figures of the
model results. Uses the Nextstrain CLI to upload figures to S3.

Currently hard-coded to only generate/upload figures for renewal model
results for GISAID Nextstrain clades at the global resolution. We can
update the workflow to use a matrix to generate/upload figures for
multiple model results when once we support visualizations for them.

I had originally planned to include this as part of the main Snakemake
pipeline. However, I ran into errors for chromium when running
the pipeline in Nextstrain docker runtime. I figured it's not worth
bloating the Nextstrain docker-base to support this when we can
accomplish the same thing in a GitHub Action workflow. This action
should be triggered by the main pipeline after is has uploaded the
latest model results to S3.