_get_all_input: ensure configs are lists
We filter on specific values for `data_provenances` and `geo_resolutions`
in the triggers for static model visualizations. This did not work as
intended when we provided override command line configs that are strings
instead of lists. This commit ensures that the `data_provenances`
and `geo_resolutions` are always lists so that they can be filtered
properly.