[viz] improve general styles
Most styles chosen to mimic nextstrain.org, and more specifically
the layout of the /sars-cov-2/forecasts/ page. This allows us to
remove the static-image-script style injection.

While most lato fonts aren't imported in the app, they are committed
here to make it simple to add them in the future as needed (./src/fonts
is ~800kb).

The changes to the chart margins revealed some dots appearing in the
top left of the (frequency) plots, which I'll investigate in a subsequent
commit.