Import viz code as a library
This signifies a shift to the ncov-specific viz code describing the
ncov-specific parts (e.g. the paragraphs, overall ordering of Panels)
and importing the JS code which performs the viz itself.

The eventual location of this ncov-specific code may not be this repo.
Since we aim to display it on nextstrain.org, that is probably the best
place to store and develop this code. There are a few moving parts
here (static-figure generation etc) to think through.

Currently using a tarball of the library as it's not yet published to
npm, which is committed here as it's only 17kb. This tarball was created
(in the `forecasts-viz` dir) via `npm pack`.