config: Remove "usa" geo-resolution
As noted in #16, the US case counts are no longer being updated and
we have not found a new source yet.

I did _not_ realize that this would break the renewal model runs, which
then causes the whole model pipeline to fail early. This commit updates
the config to remove the "usa" geo-resolution completely so our model
runs for the global data.