Swap to "seq_counts"
Previously, the prepare data script referred to "clades" being the tuple of location, clade, date and sequence count. This input file was all of the count data. It was then ran through prepare-data.py to generate "variants" consisting of the same data structure but generally reduced to a subset of locations and a subset of dates. Also, some clades are collapsed into "other".

However, I found this terminology confusing. Especially if we want to have Pango lineages as an input. I've opted to switch everything to "sequence counts", which is the term Marlin and I have been using. Here, we now have input sequence counts and prepared sequence counts that are then passed on to the models.