Run for Pango lineages and drop renewal
This uses previous commits logic to slot in `pango_lineages` for `variant_classification` that will grab sequence counts for lineages instead of clades. Other parts of the workflow including prepare-data.py should run as normal, except the function to map "clade-to-variant" fails with Pango lineages. This is because all clades that were meant to be kept had to be in the clade_to_variant.tsv file. At this point, there's no reason to do this additional mapping and I just dropped this file.

This commit also drops renewal_config from the config to stop renewal model from running. I did this because renewal model with >50 variants takes a while to run and because we need to update renewal config to allow a default generation time without specifying it for each variant. evofr should be just fine with this.