Configure pivot separately for clades and lineages
Previously, `pivot` was specified in `mlr-config.yaml` and so would be fixed across different prepared datasets. This commit adds a `models` entry to the overall `config.yaml` and populates this with `22B (Omicron)` for `nextstrain_clades` and with the equivalent `BA.5` for `pango_lineages`. 

Workflow rules in `models.smk` are used to pass in `--pivot` to `run-mlr-model.py` and the run script uses this passed in value to override whatever is included in the baseline `mlr-config.yaml`.