Fix trigger for static images
Since we stopped running the renewal models in https://github.com/nextstrain/forecasts-ncov/pull/34,
the check that all required models are included fails and the static
image workflow has not been running in our automated workflows. The
last automated run of the static image workflow was 2023-04-23.¹

This commit removes the extra check for required models since the MLR
and renewal data are now handled separately in the viz components.²
This also adds a new function to generate the inputs dynamically for the
rule `trigger_static_model_viz`. This ensures that we wait on results
from both models if we decide to run the renewal models again.

¹ https://github.com/nextstrain/forecasts-ncov/actions/runs/4781440942
² https://github.com/nextstrain/forecasts-viz/pull/3