Reduce verbosity of logging in collapse script
I found the initial printing of the entirety of low count lineages at each depth iteration to be overkill and made it hard to read what was going on. This a simple commit to update printing to just output the low count lineages specific to each depth level. I believe this makes it easier to follow what the script is doing.

I also slightly revised the top-matter to make pango_aliasor revisions more clear.