Remove static image generation steps/workflow
We've transitioned to using an iFrame for the SARS-CoV-2 forecasts
page in nextstrain.org¹ so we no longer need to generate the static
images as part of the automated workflow.

I still left the static-images.js script since it is a nice way to
generate static images of the viz.

¹ https://github.com/nextstrain/nextstrain.org/pull/691