Update pivot to 23A (XBB.1.5)
BA.5 was dropping from the Pango lineages analysis due to collapse lineage counts (leaving BA.5.1 and BA.5.2). This was causing problems with the pivot being absent. And only a short matter of time before it disappears from clades as well.

In updating to a new pivot we need a clade and a corresponding lineage that appears across geographies. XBB (which might be more natural) didn't work for this reason (basically absent from Canada).

I chose clade 23A (lineage XBB.1.5) as it's well represented across countries and should give us a fair amount of time before it needs updating again.