Use default pool_scale
Moving back to default pool_scale of 0.1. We expect the lineages-level analysis to have empirically similar behavior for pool_scale of 0.1 and 10 because this is the prior before variation in growth advantages is observed. In the lineages analysis there are a large number of lineages and from this the distribution of growth advantages is apparent. 

Behavior is as expected.