Reduce sequence count thresholds
With hierarchical model in place, we can now include locations with fewer sequences. This is particularly helpful with the hierarchical model to estimate growth advantage of emerging lineages. This commit drops count threshold for clades to 100 sequences in the previous 150 days and for lineages to 300 sequences in the previous 150 days. This results in 16 locations for clades and 10 locations for lineages.