ingest: Create Snakemake workflow for sequence counts
Creates a Snakemake workflow for the sequence counts jobs that have been
run directly GitHub Action workflows. Copies the exact options and files
used in the existing GH Action workflows (gisaid¹ and open²).

This will make it easier to run the workflows locally when needed and to
be able to use the shared pathogen-repo-build workflow³ to set up
the GitHub Actions.

¹ https://github.com/nextstrain/forecasts-ncov/blob/4fab4c912745a362c006a2bf893bf4530e050af0/.github/workflows/update-ncov-gisaid-clade-counts.yaml
² https://github.com/nextstrain/forecasts-ncov/blob/4fab4c912745a362c006a2bf893bf4530e050af0/.github/workflows/update-ncov-open-clade-counts.yaml
³ https://github.com/nextstrain/.github/blob/98a963c4dfc2be923d221349bb62f32ca26531f3/.github/workflows/pathogen-repo-build.yaml