run-models workflow: Refactor to use pathogen-repo-build
Since the workflow config has been updated in the previous commit to use
a top level `s3_dst` param, we can now easily override it in the
GH Action workflow with input values for trial runs. This allows us to
directly use the reusable pathogen-repo-build workflow.

Note that `DATA_PROVENANCES` and `S3_DST` are passed as envvars to the
reusable workflow so they can be interpolated in the build command, but
they do _not_ need to be passed to the build runtime.

Removed `bin/set-trial-optional-config` since we are no longer
using a full config file to override the workflow configs.