Post processing sets clade colours, display names
Over time setting the clade colours & display names separately to the
lineage colour definitions resulted in the two falling out of sync. This
work unifies the approach so that the same script (using a unified
configuration) modifies each JSON respectively.

Including the colour and display name data within the clades JSON is a
big improvement as it allows simple retrospective analysis of old
datasets.