modify-lineage-colours-and-order: Add default clade color
Remove the assertion that has been blocking the pipeline due to lack
of clade definition for 21L/BA.2.86.¹

Adds a default dark grey color for clades that lack clade definition so
that we can see them on the forecasts site. In the future, we may want
to add a custom Slack notification fo these clades so that we get a
louder internal notification.

¹ https://bedfordlab.slack.com/archives/C036L7V615F/p1701195514961829