ingest: restore CloudFront invalidation in upload
The CloudFront domain arg for triggering CloudFront invalidation after
upload was missed in my refactor of the GitHub Action workflow to a
Snakemake workflow in 132355457c80ad9c0c5b9e5249e869e9b3485da9.

This restores the CloudFront invalidation so that the model runs will
be able to always download the latest files.

Thank you @corneliusroemer for flagging the issue¹

¹ https://bedfordlab.slack.com/archives/C036L7V615F/p1705508089642509