Stream metadata from S3, filter, and compress
Instead of downloading the complete metadata files from S3,
uncompressing them to disk, filtering the uncompressed file to the
subset of columns we want into an uncompressed file, and deleting the
full metadata, this commit proposes streaming the original metadata from
S3 through the filter step and writing the subset of metadata to a
compressed file.

This change replaces the vendored `download-from-s3` script with `aws s3
cp` since the latter can stream to stdout while the former cannot
because it also writes log messages to stdout.

Testing with the "open" ncov dataset, this change avoids storing an 11GB
full metadata file locally and produces a 71MB compressed subset
metadata file in roughly the same amount of processing time.