Reduce location count thresholds
This commit drops location count threshold (ie the number of sequences collected in the past 30 days) from 100 to 50 for clade-level analysis and from 300 to 150 for lineage-level analysis.

With current data this goes from 8 locations included for clades to 11 locations included.

With current data this goes from 5 locations included for lineages to 7 locations included.

To support these thresholds, I looked at location count for different countries analyzed in https://bedford.io/papers/abousamra-ncov-forecasting-fit/ to get specific count thresholds. We see:
- Trinidad and Tobago with 2.3k sequences collected in 2022 and a median 30-day sequence count of 43 with a mean absolute forecasting error of 12%
- Vietnam with 6k sequences collected in 2022 and a median 30-day sequence count of 30 with a mean absolute forecasting error of 11%
- South Africa with 16k sequences collected in 2022 and a median 30-day sequence count of 170 with a mean absolute forecasting error of 7%

I believe this suggests that a threshold of 50 sequences in previous 30 days should be roughly consistent with a ~10% forecasting error. This seems like an okay threshold for public display.

It's less certain what count threshold to use for lineages where we have significantly larger number of labels than we do for clades. Keeping a 3x ratio here for now.