Update location threshold
This commit updates the location threshold for both clades and Pango lineages to be 1000 sequences from a location in the previous 150 days. The move from 30 days to 150 days makes it so that the location inclusion is based on the full analysis window. This is nicely simplifying.

I also removed the previous prune_seq_days. This had existed because exciting variants were previously submitted to databases faster. This is no longer the case.

I also dropped the minimum number of sequences to break out a clade from 2000 to 500.

I've updated the text on the viz site correspondingly