Include a location x variant count threshold for inclusion of GA
Previously, the growth advantage was included for each location and each variant regardless of the amount of data available for this particular location x variant combination. There would be situations where there is 0 sequences for a particular variant in a particular location and we'd still display the growth advantage (which was just the hierarchical prior).

This commit adds an option of "--location-ga-inclusion-threshold" to "run-mlr-model.py". I've initially set this to 100, which seemed to work well at least heuristically.