Add prepare data config files as rule inputs
Adds an input to the rule for preparing clade data that builds a list of
files required by a given model's config parameters. This list of file
paths allows the rule for preparing clade data to check for the
existence of these required files and either fail when they are missing
or run rules in the workflow to produce missing files.