Force-include clades by file path not CLI arg
Changes the interface and behavior of the prepare data script's
`--force-include-clade` argument such that the input is expected to be a
file path with a list of clade/variant pairs instead of a
space-delimited list of those pairs. This change makes the
configuration of force-included clades align with the approach to
exclude specific locations from models and also allows the workflow to
refer to files that exist in the repository already or that can be
dynamically created by workflow rules.