Finish automating push to fork of hub repo
Adds Git author configuration and GitHub personal access token to the
rule for pushing hub submissions to the fork of the hub repo. The use of
the Snakemake `envvars` declaration causes the workflow to error when
any one of those environment variables are not defined.

When running the workflow locally, each user needs to define these
variables in their own envdir. The GITHUB_TOKEN is a personal access
token with write permissions to the variant-nowcast-hub. When running
the workflow from a GitHub action, we will need to use the Nextstrain
bot's name, email, and token to automate the push.