Stub GHA to run USA models
Adds initial GitHub Action to run USA models, prepare a parquet file of
the results, and push that file to our fork of the variant nowcast hub
repo. The only trigger for this action is the manual workflow
dispatch (which won't work until this commit is merged into the default
branch). This action depends on secrets related to Git author email and
name that don't exist yet.