Remove option to push hub files to fork locally
Removes the option to push hub files to the fork of the hub repo when
running the workflow locally. Pushing to the repo requires
authentication which cannot be done safely without git credential
management software. Since the Nextstrain Docker image lacks this
software, we can't test this step of the workflow locally with the Docker
image. Instead, we'll need to test the final step of the workflow from
the corresponding GitHub Action where authentication is managed
automatically with the nextstrain-bot user.