Finish GitHub Action to upload model results
Updates the original stubbed GitHub Action workflow to actually run the
USA-specific models and upload the resulting files to a fork of the
variant hub repository. To get this working properly without exposing
GitHub tokens through shell commands, I ended up moving the git logic
out of the Snakemake workflow and into GitHub Actions. This approach
allowed me to use GHA's automatic permissions management (e.g., passing
tokens from secrets for cloning and pushing repos) to safely enable
writing to a different repository.

For this approach to work, I needed to run the models in one job with
the Docker Nextstrain runtime instead of the AWS Batch runtime, so I
could save the parquet file outputs as artifacts for the second job. The
first job creates a list of model files to be uploaded and the second
job uses that list to add those files into the forked repo.